Optimized pagerank by moving scale and shift operations into a single function, `axpb`

+++ Parse/Parse.cpp
@@ -19,7 +19,7 @@ csc_t parse_dir (const std::string dir_path){
 
     std::ifstream val, non_zero, col_idx;
 
-    std::vector<double> csc_val;
+    std::vector<float> csc_val;
     std::vector<int> csc_non_zero;
     std::vector<int> csc_col_idx;
 
@@ -40,7 +40,7 @@ csc_t parse_dir (const std::string dir_path){
 
 
     int tmp2;
-    double tmp1;
+    float tmp1;
 
     while (val >> tmp1) {
         csc_val.push_back(tmp1);
